app/requirements.txt
# Core framework
flask~=3.1.0

# Database stack
sqlalchemy~=2.0.29
psycopg2-binary~=2.9.6
alembic~=1.13.1

# Production server
gunicorn~=23.0.0
gevent~=24.11.1
python-dotenv~=1.0.1

# Development and testing tools
pytest~=8.0.0
pytest-flask~=1.3.0
pytest-cov~=5.0.0
pytest-mock~=3.12.0 